# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fortexa Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, suspended, banned]
        - in: query
          name: limit
          schema:
            type: number
            default: 50
        - in: query
          name: offset
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    put:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user info
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban a user account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/reactivate:
    post:
      operationId: reactivateUser
      tags: [admin]
      summary: Reactivate a user account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/funds:
    post:
      operationId: adjustUserFunds
      tags: [admin]
      summary: Add or remove funds from a user account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundsAdjustment"
      responses:
        "200":
          description: Funds adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/deposits:
    get:
      operationId: getAdminDeposits
      tags: [admin]
      summary: List all deposit requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: limit
          schema:
            type: number
            default: 50
        - in: query
          name: offset
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: Deposits list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransactionList"
    post:
      operationId: createAdminDeposit
      tags: [admin]
      summary: Manually add a deposit for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualDepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/deposits/{id}:
    put:
      operationId: updateAdminDeposit
      tags: [admin]
      summary: Approve, reject, or modify a deposit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /admin/announcements:
    get:
      operationId: getAdminAnnouncements
      tags: [admin]
      summary: List all announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [admin]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /admin/announcements/{id}:
    put:
      operationId: updateAnnouncement
      tags: [admin]
      summary: Update an announcement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [admin]
      summary: Delete an announcement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── Auth ──
...
    Transaction:
      type: object
      required: [id, userId, type, amount, fee, netAmount, status, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        type:
          type: string
          enum: [deposit, withdrawal, commission, gain]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        description:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Announcement:
      type: object
      required: [id, title, message, isActive, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        message:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        isActive:
          type: boolean

    # ── Admin ──
...
    AdminUserList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSummary"
        total:
          type: number
    AdminUserSummary:
      type: object
      required: [id, name, phone, email, investmentBalance, gainBalance, status, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        investmentBalance:
          type: number
        gainBalance:
          type: number
        status:
          type: string
          enum: [active, suspended, banned]
        role:
          type: string
          enum: [user, admin]
        referralCode:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminUserDetail:
      type: object
      required: [id, name, phone, email, investmentBalance, gainBalance, status, role, referralCode, createdAt, totalDeposited, totalWithdrawn, referralEarnings]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        investmentBalance:
          type: number
        gainBalance:
          type: number
        status:
          type: string
          enum: [active, suspended, banned]
        role:
          type: string
          enum: [user, admin]
        referralCode:
          type: string
        referredById:
          type: ["number", "null"]
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
        referralEarnings:
          type: number
        createdAt:
          type: string
          format: date-time
    AdminUserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
    FundsAdjustment:
      type: object
      required: [type, walletType, amount]
      properties:
        type:
          type: string
          enum: [add, subtract]
        walletType:
          type: string
          enum: [investment, gains]
        amount:
          type: number
          minimum: 0
        reason:
          type: ["string", "null"]
    AdminTransactionList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminTransaction"
        total:
          type: number
    AdminTransaction:
      type: object
      required: [id, userId, userName, userEmail, type, amount, fee, netAmount, status, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        userName:
          type: string
        userEmail:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, commission, gain]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        description:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TransactionStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        amount:
          type: ["number", "null"]
        rejectionReason:
          type: ["string", "null"]
    ManualDepositInput:
      type: object
      required: [userId, amount]
      properties:
        userId:
          type: number
        amount:
          type: number
          minimum: 0
        description:
          type: ["string", "null"]
...