# lib/api-zod/src/generated/types/adminUserSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fortexa Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserSummaryRole } from './adminUserSummaryRole';
import type { AdminUserSummaryStatus } from './adminUserSummaryStatus';

export interface AdminUserSummary {
  id: number;
  name: string;
  phone: string;
  email: string;
  investmentBalance: number;
  gainBalance: number;
  status: AdminUserSummaryStatus;
  role: AdminUserSummaryRole;
  referralCode?: string;
  createdAt: Date;
}